Small cleanup to system config and team setup.

+++ Assets/ECS Frenzy/Scripts/Components/Team.cs
@@ -5,6 +5,6 @@ namespace ECSFrenzy.Components {
   [Serializable]
   [GenerateAuthoringComponent]
   public struct Team : IComponentData {
-    public ushort Value;
+    public int Value;
   }
 }
