+++ Assets/ECS Frenzy/Scripts/Shared Components/SharedTeam.cs
@@ -5,6 +5,6 @@ namespace ECSFrenzy.SharedComponents {
   [Serializable]
   [GenerateAuthoringComponent]
   public struct SharedTeam : ISharedComponentData {
-    public ushort Value;
+    public int Value;
   }
 }
